refactored all errors for RuleError in modules

+++ modules/news/news.service.js
@@ -22,6 +22,10 @@ const {
   HISTORY_OBJ_KEYS: { AUTHOR, LANGUAGES, TITLE, TEXT },
 } = require('../../consts/history-obj-keys');
 const { transliterate } = require('../helper-functions');
+const {
+  STATUS_CODES: { NOT_FOUND, BAD_REQUEST },
+} = require('../../consts/status-codes');
+const RuleError = require('../../errors/rule.error');
 
 class NewsService {
   async getAllNews({ skip, limit, filter }) {
@@ -55,13 +59,13 @@ class NewsService {
     if (foundNews) {
       return foundNews;
     }
-    throw new Error(NEWS_NOT_FOUND);
+    throw new RuleError(NEWS_NOT_FOUND, NOT_FOUND);
   }
 
   async updateNews(id, news, upload, { _id: adminId }) {
     const foundNews = await News.findById(id).exec();
     if (!foundNews) {
-      throw new Error(NEWS_NOT_FOUND);
+      throw new RuleError(NEWS_NOT_FOUND, NOT_FOUND);
     }
 
     if (upload.length) {
@@ -76,7 +80,7 @@ class NewsService {
     }
 
     if (await this.checkNewsExist(news, id)) {
-      throw new Error(NEWS_ALREADY_EXIST);
+      throw new RuleError(NEWS_ALREADY_EXIST, BAD_REQUEST);
     }
     if (news) {
       const { beforeChanges, afterChanges } = getChanges(foundNews, news);
@@ -99,12 +103,12 @@ class NewsService {
 
   async addNews(data, upload, { _id: adminId }) {
     if (await this.checkNewsExist(data)) {
-      throw new Error(NEWS_ALREADY_EXIST);
+      throw new RuleError(NEWS_ALREADY_EXIST, BAD_REQUEST);
     }
 
     if (upload.length) {
       if (!upload[0] && !upload[1]) {
-        throw new Error(PHOTO_NOT_FOUND);
+        throw new RuleError(PHOTO_NOT_FOUND, NOT_FOUND);
       }
     }
 
@@ -132,6 +136,7 @@ class NewsService {
 
   async deleteNews(id, { _id: adminId }) {
     const foundNews = await News.findByIdAndDelete(id).exec();
+    if (!foundNews) throw new RuleError(NEWS_NOT_FOUND, NOT_FOUND);
     await uploadService.deleteFiles([foundNews.author.image, foundNews.image]);
 
     if (foundNews) {
@@ -149,7 +154,7 @@ class NewsService {
 
       return foundNews;
     }
-    throw new Error(NEWS_NOT_FOUND);
+    throw new RuleError(NEWS_NOT_FOUND, NOT_FOUND);
   }
 
   async checkNewsExist(data, id) {
